requirements.txt
python-pptx
pptxtopdf
openai
pypdf